Split hivemind.client into hivemind.averaging and hivemind.moe
