+++ tests/test_dht_schema.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel, StrictInt, conint
 from typing import Dict
 
 import hivemind
-from hivemind.dht import get_dht_time
+from hivemind.utils.timed_storage import get_dht_time
 from hivemind.dht.node import DHTNode, LOCALHOST
 from hivemind.dht.schema import BytesWithPublicKey, SchemaValidator
 from hivemind.dht.validation import DHTRecord, RecordValidatorBase
